+++ src/qdialogprogress.h
@@ -41,9 +41,9 @@ public:
 
 protected slots:
    void onButtonClicked( QAbstractButton* button );
-   void onPrepareProgress( long current, long final );
-   void onLabelProgress( long current, long final );
-   void onTimeEstimate( long time );
+   void onPrepareProgress( int current, int final );
+   void onLabelProgress( int current, int final );
+   void onTimeEstimate( int time );
    void onFinished(int status );
    void onTimeout();
